+++ plone/app/event/dx/behaviors.py
@@ -32,6 +32,7 @@ from z3c.form.browser.text import TextFieldWidget
 from z3c.form.browser.textlines import TextLinesFieldWidget
 from zope import schema
 from zope.component import adapter
+from zope.globalrequest import getRequest
 from zope.interface import alsoProvides
 from zope.interface import implementer
 from zope.interface import Invalid
@@ -319,21 +320,6 @@ alsoProvides(IEventAttendees, IFormFieldProvider)
 alsoProvides(IEventContact, IFormFieldProvider)
 
 
-"""
-if not obj.sync_uid:
-    # sync_uid has to be set for icalendar data exchange.
-    uid = IUUID(obj)
-    # We don't want to fail when getRequest() returns None, e.g when
-    # creating an event during test layer setup time.
-    request = getRequest() or {}
-    domain = request.get('HTTP_HOST')
-    obj.sync_uid = '%s%s' % (
-        uid,
-        '@%s' % domain if domain else ''
-    )
-"""
-
-
 # Attribute indexer
 
 # Start indexer
@@ -367,10 +353,10 @@ def location_indexer(obj):
 # icalendar event UID indexer
 @indexer(IDXEvent)
 def sync_uid_indexer(obj):
-    event = IEventBasic(obj)
-    if not event.sync_uid:
+    sync_uid = IEventAccessor(obj).sync_uid
+    if not sync_uid:
         raise AttributeError
-    return event.sync_uid
+    return sync_uid
 
 
 # Body text indexing
@@ -489,7 +475,7 @@ class EventAccessor(object):
 
     @start.setter
     def start(self, value):
-        return setattr(self, 'start', value)
+        return setattr(self, 'start', pydt(value))
 
     @property
     def end(self):
@@ -506,7 +492,7 @@ class EventAccessor(object):
 
     @end.setter
     def end(self, value):
-        return setattr(self, 'end', value)
+        return setattr(self, 'end', pydt(value))
 
     @property
     def timezone(self):
@@ -523,7 +509,20 @@ class EventAccessor(object):
             tz_end = tz.zone
         return tz_start if tz_start == tz_end else (tz_start, tz_end)
 
-    # rw properties not in behaviors (yet) # TODO revisit
+    @property
+    def sync_uid(self):
+        # Return externally set sync_uid or Plone's UUID + @domain.
+        sync_uid = getattr(self.context, 'sync_uid', None)
+        if not sync_uid:
+            # Return internal sync_uid
+            request = getRequest() or {}
+            domain = request.get('HTTP_HOST', None)
+            domain = '@' + domain if domain else ''
+            sync_uid = self.uid + domain if self.uid else None
+        return sync_uid
+
+    # rw properties not in behaviors.
+    # TODO: revisit, deprecate.
 
     @property
     def title(self):