+++ plone/app/event/ical/exporter.py
@@ -1,11 +1,10 @@
 from Acquisition import aq_inner
-from Products.ZCatalog.interfaces import ICatalogBrain
 from datetime import datetime
 from datetime import timedelta
 from plone.app.contentlisting.interfaces import IContentListingObject
-from plone.app.event.base import RET_MODE_BRAINS
 from plone.app.event.base import default_timezone
 from plone.app.event.base import get_events
+from plone.app.event.base import RET_MODE_BRAINS
 from plone.event.interfaces import IEvent
 from plone.event.interfaces import IEventAccessor
 from plone.event.interfaces import IICalendar
@@ -14,6 +13,7 @@ from plone.event.interfaces import IOccurrence
 from plone.event.utils import is_datetime
 from plone.event.utils import tzdel
 from plone.event.utils import utc
+from Products.ZCatalog.interfaces import ICatalogBrain
 from zope.interface import implementer
 from zope.publisher.browser import BrowserView
 
@@ -219,11 +219,7 @@ class ICalendarEventComponent(object):
 
     @property
     def uid(self):
-        if self.event.sync_uid:
-            # Re-Use existing icalendar event UID
-            return {'value': self.event.sync_uid}
-        # Else, use plone.uuid
-        return {'value': self.event.uid}
+        return {'value': self.event.sync_uid}
 
     @property
     def url(self):